March 16 updates

+++ addons/js_composer/include/classes/editors/popups/class-vc-templates-panel-editor.php
@@ -7,7 +7,7 @@ if ( ! defined( 'ABSPATH' ) ) {
  * Class Vc_Templates_Panel_Editor
  * @since 4.4
  */
-class Vc_Templates_Panel_Editor implements Vc_Render {
+class Vc_Templates_Panel_Editor {
 	/**
 	 * @since 4.4
 	 * @var string
@@ -54,7 +54,10 @@ class Vc_Templates_Panel_Editor implements Vc_Render {
 		 *  'vc_frontend_load_template' -> loading template content for frontend
 		 *  'vc_delete_template' -> deleting template by index
 		 */
-		add_action( 'wp_ajax_vc_save_template', array( &$this, 'save' ) );
+		add_action( 'wp_ajax_vc_save_template', array(
+			&$this,
+			'save',
+		) );
 		add_action( 'wp_ajax_vc_backend_load_template', array(
 			&$this,
 			'renderBackendTemplate',
@@ -67,12 +70,15 @@ class Vc_Templates_Panel_Editor implements Vc_Render {
 			&$this,
 			'renderTemplatePreview',
 		) );
-		add_action( 'wp_ajax_vc_delete_template', array( &$this, 'delete' ) );
-
-/*		add_action( 'vc-render-templates-preview-template', array(
+		add_action( 'wp_ajax_vc_delete_template', array(
 			&$this,
-			'addScriptsToTemplatePreview',
-		) );*/
+			'delete',
+		) );
+
+		/*		add_action( 'vc-render-templates-preview-template', array(
+					&$this,
+					'addScriptsToTemplatePreview',
+				) );*/
 
 	}
 
@@ -84,11 +90,7 @@ class Vc_Templates_Panel_Editor implements Vc_Render {
 		if ( 'my_templates' === $category['category'] ) {
 			$category['output'] = '';
 
-			if ( vc_user_access()
-				->part( 'templates' )
-				->checkStateAny( true, null )
-				->get()
-			) {
+			if ( vc_user_access()->part( 'templates' )->checkStateAny( true, null )->get() ) {
 				$category['output'] .= '
 				<div class="vc_column vc_col-sm-12" data-vc-hide-on-search="true">
 					<div class="vc_element_label">' . esc_html( __( 'Save current layout as a template', 'js_composer' ) ) . '</div>
@@ -186,15 +188,11 @@ class Vc_Templates_Panel_Editor implements Vc_Render {
 		$preview_template_title = esc_attr( 'Preview template', 'js_composer' );
 		$add_template_title = esc_attr( 'Add template', 'js_composer' );
 		$deleteControlRender = '';
-		if ( vc_user_access()
-			->part( 'templates' )
-			->checkStateAny( true, null )
-			->get()
-		) {
+		if ( vc_user_access()->part( 'templates' )->checkStateAny( true, null )->get() ) {
 			$delete_template_title = esc_attr( 'Delete template', 'js_composer' );
 			$deleteControlRender = <<<DATA
 				<button type="button" class="vc_general vc_ui-control-button" data-vc-ui-delete="template-title" title="$delete_template_title">
-					<i class="vc_ui-icon-pixel vc_ui-icon-pixel-control-trash-dark"></i>
+					<i class="vc-composer-icon vc-c-icon-delete_empty"></i>
 				</button>
 DATA;
 		}
@@ -205,12 +203,12 @@ DATA;
 			<div class="vc_ui-list-bar-item-actions">
 				<button type="button" class="vc_general vc_ui-control-button" title="$add_template_title"
 					 	data-template-handler="">
-					<i class="vc_ui-icon-pixel vc_ui-icon-pixel-control-add-dark"></i>
+					<i class="vc-composer-icon vc-c-icon-add"></i>
 				</button>
 				$deleteControlRender
 				<button type="button" class="vc_general vc_ui-control-button" title="$preview_template_title"
 					data-vc-container=".vc_ui-list-bar" data-vc-preview-handler data-vc-target="[data-template_id_hash=$template_id_hash]">
-					<i class="vc_ui-icon-pixel vc_ui-preview-icon"></i>
+					<i class="vc-composer-icon vc-c-icon-arrow_drop_down"></i>
 				</button>
 			</div>
 HTML;
@@ -241,11 +239,11 @@ HTML;
 		<div class="vc_ui-list-bar-item-actions">
 			<button type="button" class="vc_general vc_ui-control-button" title="$add_template_title"
 					data-template-handler="">
-				<i class="vc_ui-icon-pixel vc_ui-icon-pixel-control-add-dark"></i>
+				<i class="vc-composer-icon vc-c-icon-add"></i>
 			</button>
 			<button type="button" class="vc_general vc_ui-control-button" title="$preview_template_title"
 				data-vc-container=".vc_ui-list-bar" data-vc-preview-handler data-vc-target="[data-template_id_hash=$template_id_hash]">
-				<i class="vc_ui-icon-pixel vc_ui-preview-icon"></i>
+				<i class="vc-composer-icon vc-c-icon-arrow_drop_down"></i>
 			</button>
 		</div>
 HTML;
@@ -258,14 +256,7 @@ HTML;
 	 * vc_filter: vc_templates_render_frontend_template - called when unknown template received to render in frontend.
 	 */
 	function renderFrontendTemplate() {
-		vc_user_access()
-			->checkAdminNonce()
-			->validateDie()
-			->wpAny( 'edit_posts', 'edit_pages' )
-			->validateDie()
-			->part( 'templates' )
-			->can()
-			->validateDie();
+		vc_user_access()->checkAdminNonce()->validateDie()->wpAny( 'edit_posts', 'edit_pages' )->validateDie()->part( 'templates' )->can()->validateDie();
 
 		add_filter( 'vc_frontend_template_the_content', array(
 			&$this,
@@ -319,19 +310,6 @@ HTML;
 	}
 
 	/**
-	 * @since 4.4
-	 * @deprecated 4.7
-	 *
-	 * vc_filter: vc_templates_render_template - hook to override singe template rendering in panel window
-	 */
-	public function render() {
-		_deprecated_function( '\Vc_Templates_Panel_Editor::render', '4.7 (will be removed in 4.11)', '\Vc_Templates_Panel_Editor::renderUITemplate' );
-		vc_include_template( 'editors/popups/panel_templates.tpl.php', array(
-			'box' => $this,
-		) );
-	}
-
-	/**
 	 * @since 4.7
 	 */
 	public function renderUITemplate() {
@@ -346,14 +324,7 @@ HTML;
 	 * @since 4.4
 	 */
 	public function save() {
-		vc_user_access()
-			->checkAdminNonce()
-			->validateDie()
-			->wpAny( 'edit_posts', 'edit_pages' )
-			->validateDie()
-			->part( 'templates' )
-			->checkStateAny( true, null )
-			->validateDie();
+		vc_user_access()->checkAdminNonce()->validateDie()->wpAny( 'edit_posts', 'edit_pages' )->validateDie()->part( 'templates' )->checkStateAny( true, null )->validateDie();
 
 		$template_name = vc_post_param( 'template_name' );
 		$template = vc_post_param( 'template' );
@@ -395,14 +366,7 @@ HTML;
 	 * vc_filter: vc_templates_render_backend_template - called when unknown template requested to render in backend
 	 */
 	public function renderBackendTemplate() {
-		vc_user_access()
-			->checkAdminNonce()
-			->validateDie()
-			->wpAny( 'edit_posts', 'edit_pages' )
-			->validateDie()
-			->part( 'templates' )
-			->can()
-			->validateDie();
+		vc_user_access()->checkAdminNonce()->validateDie()->wpAny( 'edit_posts', 'edit_pages' )->validateDie()->part( 'templates' )->can()->validateDie();
 
 		$template_id = vc_post_param( 'template_unique_id' );
 		$template_type = vc_post_param( 'template_type' );
@@ -437,14 +401,10 @@ HTML;
 	 * @since 4.8
 	 */
 	public function renderTemplatePreview() {
-		vc_user_access()
-			->checkAdminNonce()
-			->validateDie()
-			->wpAny( array( 'edit_post', (int) vc_request_param( 'post_id' ) ) )
-			->validateDie()
-			->part( 'templates' )
-			->can()
-			->validateDie();
+		vc_user_access()->checkAdminNonce()->validateDie()->wpAny( array(
+				'edit_post',
+				(int) vc_request_param( 'post_id' ),
+			) )->validateDie()->part( 'templates' )->can()->validateDie();
 
 		$template_id = vc_request_param( 'template_unique_id' );
 		$template_type = vc_request_param( 'template_type' );
@@ -478,6 +438,7 @@ HTML;
 		die();
 
 	}
+
 	public function registerPreviewScripts() {
 		visual_composer()->registerAdminJavascript();
 		visual_composer()->registerAdminCss();
@@ -487,6 +448,7 @@ HTML;
 			'vc-backend-min-js',
 		), WPB_VC_VERSION, true );
 	}
+
 	/**
 	 * Enqueue required scripts for template preview
 	 * @since 4.8
@@ -501,29 +463,28 @@ HTML;
 	 * @since 4.4
 	 */
 	public function delete() {
-		vc_user_access()
-			->checkAdminNonce()
-			->validateDie()
-			->wpAny( 'edit_posts', 'edit_pages' )
-			->validateDie()
-			->part( 'templates' )
-			->checkStateAny( true, null )
-			->validateDie();
+		vc_user_access()->checkAdminNonce()->validateDie()->wpAny( 'edit_posts', 'edit_pages' )->validateDie()->part( 'templates' )->checkStateAny( true, null )->validateDie();
 
 		$template_id = vc_post_param( 'template_id' );
+		$template_type = vc_post_param( 'template_type' );
 
 		if ( ! isset( $template_id ) || '' === $template_id ) {
 			die( 'Error: Vc_Templates_Panel_Editor::delete:1' );
 		}
+		if ( 'my_templates' === $template_type ) {
 
-		$saved_templates = get_option( $this->option_name );
-		unset( $saved_templates[ $template_id ] );
-		if ( count( $saved_templates ) > 0 ) {
-			update_option( $this->option_name, $saved_templates );
+			$saved_templates = get_option( $this->option_name );
+			unset( $saved_templates[ $template_id ] );
+			if ( count( $saved_templates ) > 0 ) {
+				update_option( $this->option_name, $saved_templates );
+			} else {
+				delete_option( $this->option_name );
+			}
+			wp_send_json_success();
 		} else {
-			delete_option( $this->option_name );
+			do_action( 'vc_templates_delete_templates', $template_id, $template_type );
 		}
-		die();
+		wp_send_json_error();
 	}
 
 	/**
@@ -540,6 +501,16 @@ HTML;
 		$total_templates_to_show = apply_filters( 'vc_load_default_templates_limit_total', 6 );
 
 		return array_slice( $templates, $start_index, $total_templates_to_show );
+	}
+
+	/**
+	 * Get user templates
+	 *
+	 * @since 4.12
+	 * @return mixed
+	 */
+	public function getUserTemplates() {
+		return apply_filters( 'vc_get_user_templates', get_option( $this->option_name ) );
 	}
 
 	/**
@@ -560,7 +531,7 @@ HTML;
 		// Here we go..
 		if ( apply_filters( 'vc_show_user_templates', true ) ) {
 			// We need to get all "My Templates"
-			$user_templates = apply_filters( 'vc_get_user_templates', get_option( $this->option_name ) );
+			$user_templates = $this->getUserTemplates();
 			// this has only 'name' and 'template' key  and index 'key' is template id.
 			$arr_category = array(
 				'category' => 'my_templates',
@@ -595,6 +566,9 @@ HTML;
 			);
 			$category_templates = array();
 			foreach ( $default_templates as $template_id => $template_data ) {
+				if ( isset( $template_data['disabled'] ) && $template_data['disabled'] ) {
+					continue;
+				}
 				$category_templates[] = array(
 					'unique_id' => $template_id,
 					'name' => $template_data['name'],
@@ -605,8 +579,10 @@ HTML;
 					'custom_class' => isset( $template_data['custom_class'] ) ? $template_data['custom_class'] : false,
 				);
 			}
-			$arr_category['templates'] = $category_templates;
-			$data[] = $arr_category;
+			if ( ! empty( $category_templates ) ) {
+				$arr_category['templates'] = $category_templates;
+				$data[] = $arr_category;
+			}
 		}
 
 		// To get any other 3rd "Custom template" - do this by hook filter 'vc_get_all_templates'
@@ -689,7 +665,6 @@ HTML;
 			if ( ! is_array( $this->default_templates ) ) {
 				$this->default_templates = array();
 			}
-
 			$this->default_templates[] = $data;
 
 			return true;
@@ -729,7 +704,10 @@ HTML;
 	 */
 	public function sortTemplatesByCategories( array $data ) {
 		$buffer = $data;
-		usort( $buffer, array( &$this, 'cmpCategory' ) );
+		uasort( $buffer, array(
+			&$this,
+			'cmpCategory',
+		) );
 
 		return $buffer;
 	}
@@ -743,7 +721,10 @@ HTML;
 	 */
 	public function sortTemplatesByNameWeight( array $data ) {
 		$buffer = $data;
-		usort( $buffer, array( &$this, 'cmpNameWeight' ) );
+		uasort( $buffer, array(
+			&$this,
+			'cmpNameWeight',
+		) );
 
 		return $buffer;
 	}
@@ -839,7 +820,6 @@ HTML;
 	 * @todo move to autoload or else some where.
 	 * @since 4.4.3
 	 *
-	 * @return string
 	 */
 	public function addFrontendTemplatesShortcodesCustomCss() {
 		$output = $shortcodes_custom_css = '';
@@ -886,5 +866,9 @@ HTML;
 HTML;
 
 		return $output;
+	}
+
+	public function getOptionName() {
+		return $this->option_name;
 	}
 }
